Bump GeckoView to v95.0.20211006094130 (#3969)
* Bump GeckoView to v95.0.20211006094130

* Implement MediaSession WebExtension

* Implement media playback shortcut in the navigation bar

* Fix bookmarks panel surface issues with layers

* Implement yaml based preferences

+++ app/src/common/shared/org/mozilla/vrbrowser/ui/widgets/NavigationBarWidget.java
@@ -802,7 +802,7 @@ public class NavigationBarWidget extends UIWidget implements GeckoSession.Naviga
         if (mFullScreenMedia != null && mFullScreenMedia.getWidth() > 0 && mFullScreenMedia.getHeight() > 0) {
             final boolean resetBorder = aProjection == VideoProjectionMenuWidget.VIDEO_PROJECTION_360 ||
                     aProjection == VideoProjectionMenuWidget.VIDEO_PROJECTION_360_STEREO;
-            mAttachedWindow.enableVRVideoMode(mFullScreenMedia.getWidth(), mFullScreenMedia.getHeight(), resetBorder);
+            mAttachedWindow.enableVRVideoMode((int)mFullScreenMedia.getWidth(), (int)mFullScreenMedia.getHeight(), resetBorder);
             // Handle video resize while in VR video playback
             mFullScreenMedia.setResizeDelegate((width, height) -> {
                 mAttachedWindow.enableVRVideoMode(width, height, resetBorder);
